fix: apply 50+ audit fixes across settlement, materials, VFX, security, interiors
- Settlement generator: room_functions, building floors, cluster maps, lot clipping
- Materials: cobblestone key fix, glass/water recipes, bronze patina, dead Musgrave
- VFX templates: sanitize_cs_identifier fixes, PrimeTween guards, deprecated casts
- Security: remove type from BLOCKED_FUNCTIONS per user policy
- Interior binding: add 9 Hearthvale building types to BUILDING_ROOM_MAP
- Tool runner: fix deprecated tempfile.mktemp, add logger
- Unity templates: fix IDamageable, projectile damage, monster avatar type
- Export: add GLTF PBR settings, animation timing sidecar
- 293 new tests added across 11 test files

+++ Tools/mcp-toolkit/src/veilbreakers_mcp/shared/unity_templates/combat_vfx_templates.py
@@ -390,7 +390,7 @@ public class VB_ComboVFX_{safe_brand} : MonoBehaviour
         emission.enabled = true;
         emission.rateOverTime = 0f;
         emission.SetBursts(new ParticleSystem.Burst[] {{
-            new ParticleSystem.Burst(0f, (short)baseParticleCount)
+            new ParticleSystem.Burst(0f, (int)baseParticleCount)
         }});
 
         var shape = hitBurstPS.shape;
@@ -528,7 +528,7 @@ public class VB_ComboVFX_{safe_brand} : MonoBehaviour
 
         var emission = hitBurstPS.emission;
         emission.SetBursts(new ParticleSystem.Burst[] {{
-            new ParticleSystem.Burst(0f, (short)burstCount)
+            new ParticleSystem.Burst(0f, (int)burstCount)
         }});
 
         // Gradient: primary -> glow -> fade
@@ -1237,7 +1237,7 @@ public class ChannelVFXController : MonoBehaviour
         var emission = releasePS.emission;
         int burstCount = (int)(baseEmissionRate * releaseBurstMultiplier * (0.5f + intensityRatio));
         emission.SetBursts(new ParticleSystem.Burst[] {{
-            new ParticleSystem.Burst(0f, (short)Mathf.Clamp(burstCount, 20, 500))
+            new ParticleSystem.Burst(0f, (int)Mathf.Clamp(burstCount, 20, 500))
         }});
 
         var renderer = releasePS.GetComponent<ParticleSystemRenderer>();
